+++ VERSION_HISTORY.md
@@ -157,6 +157,7 @@
 * Fix delay_ms. On small ms delays it works correctly, but on seconds delay it doubled.
 * Rename pin.h to gpio.h
 * Read byte from EEPROM via IAP registers implemented
+* TODO Power management low voltage flag routines immplemented
 * TODO Read byte from EEPROM via MOVC command implemented
 * TODO Write byte to EEPROM
 * TODO Erase EEPROM segment